+++ src/core/engine/decode/decodercontext.h
@@ -62,6 +62,7 @@ public:
     [[nodiscard]] bool isValid() const;
     [[nodiscard]] bool isDecoding() const;
     [[nodiscard]] bool isSeekable() const;
+    [[nodiscard]] AudioDecoder::RepeatHandling repeatHandling() const;
     [[nodiscard]] AudioStreamPtr activeStream() const;
     [[nodiscard]] StreamId activeStreamId() const;
     [[nodiscard]] const Track& track() const;