+++ src/core/engine/decode/decodingcontroller.h
@@ -84,6 +84,7 @@ public:
     [[nodiscard]] bool isValid() const;
     [[nodiscard]] bool isDecoding() const;
     [[nodiscard]] bool isSeekable() const;
+    [[nodiscard]] AudioDecoder::RepeatHandling repeatHandling() const;
 
     [[nodiscard]] AudioStreamPtr activeStream() const;
     [[nodiscard]] StreamId activeStreamId() const;